[level 1] Title: 부족한 금액 계산하기, Time: 0.08 ms, Memory: 78 MB -BaekjoonHub

+++ 프로그래머스/lv1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -1,12 +1,12 @@
 class Solution {
-    public long solution(int price, int money, int cnt) {
+    public long solution(long price, long money, long cnt) {
         long answer = 0;
         long sum = 0;
         for(int i=1; i<=cnt; i++){
-            sum += (long) (price * i);
+            sum += price * i;
         }
         if ((sum-money) > 0){
-            answer = (long) (sum-money);
+            answer = sum-money;
         } 
 
         return answer;
